auth/requirements.txt
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.0
cffi==1.15.0
click==8.0.4
cryptography==36.0.1
fastapi==0.75.0
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.4.0
Mako==1.2.0
MarkupSafe==2.1.1
passlib==1.7.4
psycopg2==2.9.3
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
python-dotenv==0.19.2
PyYAML==6.0
sniffio==1.2.0
SQLAlchemy==1.4.32
starlette==0.17.1
typing_extensions==4.1.1
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8
websockets==10.2
zipp==3.7.0
